backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Real-time (WebSockets)
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Background Tasks
celery>=5.3.0
redis>=5.0.0

# Utils
python-decouple>=3.8
Pillow>=10.0.0
django-filter>=23.5

# Development
django-debug-toolbar>=4.2.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-asyncio>=0.21.0
